[/] Renamed specification getters from "GetSpec.." to "GetSpecs.."..
Well, original idea was that say counter 2 has _specification_:

  Counter value width (bits): 16
  Available prescalers (powers of two): ( 0 3 6 8 10 )

But what is the name of specific calculation setting?:

  Counter value width: 16
  Prescaler (power of two): 3

Is it "specification"?

.. We don't think we got right solution for this. But that's what
is done:

Record which contains list of prescalers is called "clock scaling options".
Record with prescaler is called "clock scale setting".
Calculated result record is called "clock scale".

Functions which deal with list of prescalers are named with "Specs".
Functions which deal with prescaler are named with "Spec".
Functions which deal with calculated result record has no suffix.