[^-/] [Interface] Removed datatype for counter specific limit setting..
We often need two entities for calculations: prescale and scale variable
size in bits. They are weird and stored weirdly and we decided to
make record to store them there and see them less.

But this record serves no practical purpose and clutters namespace
and dusts conception. So that record now is removed and functions
that used it changed to use two arguments.

There is another type record - counter limits record. "Limits" - plural.
That record is varsize in bits and list of prescales. "List" - several
values.

That record is useful and actually design decision.